requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.3
astor==0.8.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
black==19.10b0
bleach==3.1.3
burnysc2==4.11.15
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.14.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
more-itertools==7.2.0
mpyq==0.2.5
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.17.3
opencv-python==4.1.1.26
opt-einsum==3.2.0
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pluggy==0.13.1
ply==3.11
portpicker==1.3.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
py==1.8.1
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
regex==2020.4.4
requests==2.22.0
s2clientprotocol==4.10.3.76114.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
unsync==1.2.1
urllib3==1.25.8
virtualenv==20.0.7
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.1
yappi==1.2.3
yarl==1.4.2
zipp==3.0.0
